+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - TableViewModel for admin.latte
 - `footer.latte` populates the default latte layout `{layout '../vendor/seablast/seablast/views/BlueprintWeb.latte'}`.
+- namespace Seablast\Seablast\Admin;
 
 ### `Changed` for changes in existing functionality